Further modularization of config components
