+++ konduit-serving-models/konduit-serving-samediff-config/pom.xml
@@ -30,11 +30,6 @@
     <artifactId>konduit-serving-samediff-config</artifactId>
 
     <dependencies>
-        <dependency>
-            <groupId>ai.konduit.serving</groupId>
-            <artifactId>konduit-serving-nd4j</artifactId>
-            <version>${project.version}</version>
-        </dependency>
 
         <dependency>
             <groupId>ai.konduit.serving</groupId>
