+++ konduit-serving-vertx-config/src/main/java/ai/konduit/serving/vertx/config/InferenceDeploymentResult.java
